Drop broken feature which statically configures DHCP addresses
NetworkManager dispatcher scripts 30-static-dhcp and 30-static-dhcpv6
modify the NetworkManager connection of the OVS interface ovs-if-br-ex
in case of a infinite DHCP lease: They will change ipv4.method and
ipv6.method from auto to manual in order to statically assign DHCP
addresses etc.

The purpose of this feature was to enable nodes to keep their network
configuration and network connectivity when their DHCP server becomes
unavailable for any reason [0]. For example, this allowed users to
deploy OpenShift nodes with "static" ip addresses, i.e. ip addresses
with infitite DHCP leases, using an initial DHCP server that would then
be removed once the cluster is up.

However, the changes by the dispatcher scripts to NM connection
ovs-if-br-ex were saved to disk which conflicts with the rollback
behaviour of configure-ovs.sh (ovs-configuration.service):
OVN-Kubernetes configuration is rolled back and (re)generated on every
boot, to take in account any changes that have possibly been applied in
the standard configuration sources. In OpenShift releases without
PR #4877 [2], the NM connection file for ovs-if-br-ex in
/etc/NetworkManager/system-connections will not be removed during
rollback which causes conflicts when changes have been applied in the
standard configuration sources. With PR #4877 [2], the NM connection
file is removed during rollback, defeating the purpose of this feature.

Nowadays, OpenShift has proper support for static ip addresses built-in
and this inf-lease-to-static feature has not been documented since 4.9.
Both dispatcher scripts were only kept due to the "no breaking changes"
policy in OpenShift. However, because of the conflict with
configure-ovs.sh's rollback feature, those dispatcher scripts could
finally be removed.

[0] https://github.com/openshift/machine-config-operator/pull/2188
[1] https://github.com/openshift/machine-config-operator/pull/4954
[2] https://github.com/openshift/machine-config-operator/pull/4877